+++ packages/vue-mc/src/Structures/Model.js
@@ -6,4 +6,18 @@ export default class Model extends VueMCModel {
     createRequest(config) {
         return new Request(config, Vue.prototype.$nuxt.$axios)
     }
+
+    options() {
+        return {
+            useFirstErrorOnly: true,
+            methods: {
+                fetch: 'GET',
+                save: 'POST',
+                update: 'PUT',
+                create: 'POST',
+                patch: 'PATCH',
+                delete: 'DELETE'
+            }
+        }
+    }
 }