+++ packages/vue-mc/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.5](https://git.awescode.com/awes-io/collection/compare/@awes-io/vue-mc@0.2.4...@awes-io/vue-mc@0.2.5) (2020-01-30)
+
+
+### Bug Fixes
+
+* linter syntax ([fa9a175](https://git.awescode.com/awes-io/collection/commits/fa9a1752d298ed1d56caf8418ad5301797d20b19))
+* model methods ([496bab7](https://git.awescode.com/awes-io/collection/commits/496bab7c78008bb42b817247aed4eaca6fcdd3c6))
+
+
+
+
+
+## [0.2.4](https://git.awescode.com/awes-io/collection/compare/@awes-io/vue-mc@0.2.3...@awes-io/vue-mc@0.2.4) (2020-01-30)
+
+
+### Bug Fixes
+
+* changed name of the package ([637253b](https://git.awescode.com/awes-io/collection/commits/637253b646dd03ab2d735ed9b36d43f7b9e3e037))
+* error message for model property ([8b408dd](https://git.awescode.com/awes-io/collection/commits/8b408ddac8a0db28a86fbbf17d32f6a11c556f82))
+
+
+
+
+
 ## [0.2.3](https://git.awescode.com/awes-io/collection/compare/@awes-io/vue-mc@0.2.2...@awes-io/vue-mc@0.2.3) (2020-01-29)
 
 ### Bug Fixes